Unmarshal: disallow nil packets
Also move CompoundPacket into its own file so it has room to grow.

Relates to #13
